api/requirements.txt
######################################
# Main Applications.                 #
######################################
Django==1.11.5
djangorestframework==3.6.3
Pillow==5.0.0
psycopg2==2.7.4
django-redis==4.4.3
git+https://github.com/nxexox/django-rest-framework-auto.git


#######################################
# Tests.                              #
#######################################
pytest-django==3.1.2
mock==2.0.0
coverage==4.4.1
factory-boy==2.8.1
